Use native Anchor accounts for governance addin api (#39)
* chore: use native Anchor accounts for governance addin api

* chore: update and publish npm
