+++ programs/nft-voter/Cargo.toml
@@ -23,7 +23,6 @@ itertools = "0.10.2"
 mpl-token-metadata = { version = "1.1.0", features = ["no-entrypoint"] }
 solana-program = "1.9.5"
 spl-governance = { version = "2.2.2", features = ["no-entrypoint"] }
-spl-governance-addin-api  = "0.1.1"
 spl-governance-tools=  "0.1.2"
 spl-token = { version = "3.3", features = [ "no-entrypoint" ] }
