+++ programs/nft-voter/src/instructions/create_max_voter_weight_record.rs
@@ -2,7 +2,7 @@ use anchor_lang::prelude::*;
 use anchor_spl::token::Mint;
 use spl_governance::state::realm;
 
-use crate::state::MaxVoterWeightRecord;
+use crate::state::max_voter_weight_record::MaxVoterWeightRecord;
 
 /// Creates MaxVoterWeightRecord used by spl-gov
 /// This instruction should only be executed once per realm/governing_token_mint to create the account
@@ -43,9 +43,6 @@ pub fn create_max_voter_weight_record(ctx: Context<CreateMaxVoterWeightRecord>) 
     )?;
 
     let max_voter_weight_record = &mut ctx.accounts.max_voter_weight_record;
-
-    max_voter_weight_record.account_discriminator =
-        spl_governance_addin_api::max_voter_weight::MaxVoterWeightRecord::ACCOUNT_DISCRIMINATOR;
 
     max_voter_weight_record.realm = ctx.accounts.realm.key();
     max_voter_weight_record.governing_token_mint = ctx.accounts.realm_governing_token_mint.key();
