+++ programs/nft-voter/src/state/mod.rs
@@ -5,16 +5,11 @@ pub use collection_config::*;
 pub mod collection_config;
 
 pub use nft_vote_record::*;
-
 pub mod nft_vote_record;
 
+pub mod max_voter_weight_record;
+
+pub use voter_weight_record::*;
+pub mod voter_weight_record;
+
 pub mod idl_types;
-
-use crate::max_voter_weight_record;
-use crate::voter_weight_record;
-
-// Generate a VoterWeightRecord and MaxVoterWeightRecord Anchor wrapper, owned by the current program
-// The Records accounts  are unique in that they are defined by the SPL governance program as ABI
-// but they are actually owned by this program
-voter_weight_record!(crate::ID);
-max_voter_weight_record!(crate::ID);
