+++ programs/nft-voter/tests/update_voter_weight_record.rs
@@ -1,11 +1,11 @@
 use crate::program_test::nft_voter_test::ConfigureCollectionArgs;
 use gpl_nft_voter::error::NftVoterError;
+use gpl_nft_voter::state::*;
 use program_test::nft_voter_test::NftVoterTest;
 use program_test::token_metadata_test::CreateNftArgs;
 use program_test::tools::*;
 use solana_program_test::*;
 use solana_sdk::transport::TransportError;
-use spl_governance_addin_api::voter_weight::VoterWeightAction;
 
 mod program_test;
